pyproject.toml
[project]
name = "wifi-channel-optimizer"
version = "0.1.0"
description = "Terminal dashboard for WiFi channel monitoring, scanning, and optimization"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "pycryptodome>=3.20.0",
    "requests>=2.31.0",
    "textual>=0.61.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "keyring>=25.0.0",
    "pyobjc-framework-corewlan>=12.1",
    "textual-plotext>=1.0.1",
    "pyyaml>=6.0",
]

[project.scripts]
wifi-monitor = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["*.py", "drivers/*.py", "devices/*.yaml"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]
